rstan: getting summaries such as ESS to R

+++ rstan/rstan/inst/include/rstan/stan_args.hpp
@@ -85,10 +85,10 @@ namespace rstan {
   private:
     bool sample_file_flag; // true: write out to a file; false, do not 
     std::string sample_file; // the file for outputting the samples 
-    unsigned int iter;   // number of iterations 
-    unsigned int warmup; // number of warmup 
-    unsigned int thin; 
-    unsigned int refresh; 
+    size_t iter;   // number of iterations 
+    size_t warmup; // number of warmup 
+    size_t thin; 
+    size_t refresh; 
     int leapfrog_steps; 
     double epsilon; 
     int max_treedepth; 
@@ -145,16 +145,16 @@ namespace rstan {
 
       idx = find_index(args_names, std::string("iter")); 
       if (idx == args_names.size()) iter = 2000U;  
-      else iter = Rcpp::as<unsigned int>(in[idx]); 
+      else iter = Rcpp::as<size_t>(in[idx]); 
 
       idx = find_index(args_names, std::string("warmup")); 
       if (idx == args_names.size()) warmup = iter / 2; 
-      else warmup = Rcpp::as<unsigned int>(in[idx]); 
+      else warmup = Rcpp::as<size_t>(in[idx]); 
 
       idx = find_index(args_names, std::string("thin")); 
-      unsigned int calculated_thin = (iter - warmup) / 1000U;
+      size_t calculated_thin = (iter - warmup) / 1000U;
       if (idx == args_names.size()) thin = (calculated_thin > 1) ? calculated_thin : 1U;
-      else thin = Rcpp::as<unsigned int>(in[idx]); 
+      else thin = Rcpp::as<size_t>(in[idx]); 
 
       idx = find_index(args_names, std::string("leapfrog_steps"));
       if (idx == args_names.size()) leapfrog_steps = -1; 
@@ -186,7 +186,7 @@ namespace rstan {
       
       idx = find_index(args_names, std::string("refresh")); 
       if (idx == args_names.size())  refresh = 1; 
-      else refresh = Rcpp::as<unsigned int>(in[idx]); 
+      else refresh = Rcpp::as<size_t>(in[idx]); 
 
 
       idx = find_index(args_names, std::string("seed")); 
@@ -194,7 +194,7 @@ namespace rstan {
         random_seed = std::time(0); 
         random_seed_src = "random"; 
       } else {
-        random_seed = Rcpp::as<unsigned int>(in[idx]); 
+        random_seed = Rcpp::as<size_t>(in[idx]); 
         random_seed_src = "user"; 
       }
 
@@ -203,7 +203,7 @@ namespace rstan {
         chain_id = 1; 
         chain_id_src = "default"; 
       } else {
-        chain_id = Rcpp::as<unsigned int>(in[idx]); 
+        chain_id = Rcpp::as<size_t>(in[idx]); 
         chain_id_src = "user"; 
       }
       
@@ -231,7 +231,7 @@ namespace rstan {
     size_t get_num_chains() const {
       return num_chains; 
     } 
-    void set_random_seed(unsigned int seed) {
+    void set_random_seed(size_t seed) {
       random_seed = seed;
     } 
     const std::string& get_random_seed_src() const {
@@ -253,13 +253,13 @@ namespace rstan {
     bool get_sample_file_flag() const { 
       return sample_file_flag; 
     }
-    unsigned int get_warmup() const {
+    size_t get_warmup() const {
       return warmup; 
     } 
-    unsigned int get_refresh() const { 
+    size_t get_refresh() const { 
       return refresh; 
     } 
-    unsigned int get_thin() const {
+    size_t get_thin() const {
       return thin;
     } 
     int get_leapfrog_steps() const {
